requirement.txt
hdfs==2.7.2
duckdb==0.8.1
pandas==2.1.0
SQLAlchemy==2.0.20
Faker==19.6.1
pymongo==4.5.0
clickhouse-driver==0.2.6
pyarrow==13.0.0
pyodps==0.11.4.1
loguru==0.7.2